Add ability for GMs to self-register for tables.
This requires that the appropriate field be selected in the table creation,
which is the default.

TODO - should send an email to the GM and to the registration email address
whenever a GM is assigned to a table

TODO - need to add boilerplate explaining how to request to be removed from
a table (which will be included in the email as well)

Also removed some commented out blocks, and (accidently) reformatted some files.