Add the "tickets" to the functionality
There is both a CSV and JSON API (only the CSV is linked to)

Note that CSV is not currently sorting (since I cannot figure out
how to make it work) so the user should sort after importing to
Excel.

This should have all the fields for a batch printing.

TODO - ADD in tickets for empty seats.