Initial work for adding Attendence Policy for the In Person Event (#183)
This adds in a field to the event which will store the Attendance Policy

Doing it this way so that it is more usable for more conventions. Each can have their own policy.

Note: the policy needs to have full HTML markup, else it will just look like a blob of text. We can have all or none (currently) for markup.